extract NewTransaction from new-compile-transaction to its own file
