--- packages/transactions/src/create-transaction-message.ts
@@ -1,19 +1,0 @@
-import { Transaction, TransactionVersion } from './types';
-
-type TransactionMessageConfig<TVersion extends TransactionVersion> = Readonly<{
-    version: TVersion;
-}>;
-
-export function createTransactionMessage<TVersion extends TransactionVersion>(
-    config: TransactionMessageConfig<TVersion>,
-): Extract<Transaction, { version: TVersion }>;
-export function createTransactionMessage<TVersion extends TransactionVersion>({
-    version,
-}: TransactionMessageConfig<TVersion>): Transaction {
-    const out: Transaction = {
-        instructions: [],
-        version,
-    };
-    Object.freeze(out);
-    return out;
-}
